Make cdc consumer singleton
