+++ superduperdb/db/base/vector_task_factory.py
