+++ superduperdb/db/mongodb/cdc/base.py
@@ -1,37 +1,13 @@
 import dataclasses as dc
 import json
 import typing as t
-from abc import ABC, abstractmethod
-from enum import Enum
 
 from bson import objectid
-from pymongo.change_stream import CollectionChangeStream
 
 from superduperdb.container.serializable import Serializable
+from superduperdb.db.base.base_cdc import DBEvent, Packet
 
 TokenType = t.Dict[str, str]
-
-
-class BaseDatabaseListener(ABC):
-    """
-    A Base class which defines basic functions to implement.
-    """
-
-    @abstractmethod
-    def listen(self):
-        pass
-
-    @abstractmethod
-    def stop(self):
-        pass
-
-    @abstractmethod
-    def setup_cdc(self) -> CollectionChangeStream:
-        pass
-
-    @abstractmethod
-    def next_cdc(self, stream: CollectionChangeStream) -> None:
-        pass
 
 
 class CachedTokens:
@@ -55,14 +31,6 @@ class CachedTokens:
         return self._current_tokens
 
 
-class DBEvent(str, Enum):
-    """`DBEvent` simple enum to hold mongo basic events."""
-
-    delete = 'delete'
-    insert = 'insert'
-    update = 'update'
-
-
 class ObjectId(objectid.ObjectId):
     @classmethod
     def __get_validators__(cls):
@@ -76,7 +44,7 @@ class ObjectId(objectid.ObjectId):
 
 
 @dc.dataclass
-class Packet:
+class MongoDBPacket(Packet):
     """
     A base packet to represent message in task queue.
     """
@@ -85,20 +53,3 @@ class Packet:
     query: t.Optional[Serializable]
 
     event_type: DBEvent = DBEvent.insert
-
-    @property
-    def is_delete(self) -> bool:
-        return self.event_type == DBEvent.delete
-
-    @staticmethod
-    def collate(packets: t.Sequence['Packet']) -> 'Packet':
-        """
-        Collate a batch of packets into one
-        """
-
-        ids = [packet.ids[0] for packet in packets]
-        query = packets[0].query
-
-        # TODO: cluster Packet for each event.
-        event_type = packets[0].event_type
-        return Packet(ids=ids, query=query, event_type=event_type)
